pyproject.toml
[tool.poetry]
name = "video-downloader-bot"
version = "0.2.0"
description = "Telegram bot for downloading videos from social networks"
authors = ["Jag_k <jag-k@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.1"
aiohttp = "^3.8.3"
pytube = "^12.1.2"
python-dotenv = "^1.0.0"
pytz = "^2023.2"
aiofiles = "^23.1.0"
pydantic = "^1.10.7"
motor = "^3.1.2"
mongopersistence = "^0.3.0"

[tool.poetry.group.dev.dependencies]
babel = "^2.11.0"
black = "^22.3.0"
pre-commit = "^2.20.0"
ruff = "^0.0.257"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
include = '\.pyi?$'
force-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.ruff_cache
  | locales/.*
)/
'''

[tool.ruff]
line-length = 80
target-version = "py311"
exclude = ["*locales*"]
select = [
    "E", # pyflakes
    "F", # pycodestyle errors
    "W", # pycodestyle warnings
    "UP", # pyupgrade
    "I", # isort
    "C4", # flake8-comprehensions
    # pytest
    "PT018", # Assertion should be broken down into multiple parts
    "PT022", # No teardown in fixture {name}, use return instead of yield
]
ignore = ["E501"]
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F405", "F811"]
"__main__.py" = ["E402", "F401", "F403", "F405", "F811"]


